rename folder of SyncHub
