+++ backend/src/Designer/EventHandlers/FileSyncHandlerExecutor.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 using System.Threading.Tasks;
-using Altinn.Studio.Designer.Hubs.SyncHub;
+using Altinn.Studio.Designer.Hubs.Sync;
 using Altinn.Studio.Designer.Models;
 using Microsoft.AspNetCore.SignalR;
