+++ backend/src/Designer/EventHandlers/LanguageTextsKeyChanged/LanguageTextsIdChanged.cs
@@ -3,7 +3,7 @@ using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Altinn.Studio.Designer.Events;
-using Altinn.Studio.Designer.Hubs.SyncHub;
+using Altinn.Studio.Designer.Hubs.Sync;
 using Altinn.Studio.Designer.Services.Interfaces;
 using MediatR;
 using Microsoft.AspNetCore.SignalR;
