+++ backend/src/Designer/Hubs/Sync/ISyncClient.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace Altinn.Studio.Designer.Hubs.SyncHub;
+namespace Altinn.Studio.Designer.Hubs.Sync;
 
 public interface ISyncClient
