+++ backend/src/Designer/Hubs/Sync/Source.cs
@@ -1,3 +1,3 @@
-namespace Altinn.Studio.Designer.Hubs.SyncHub;
+namespace Altinn.Studio.Designer.Hubs.Sync;
 
 public record Source(string Name, string Path);
