+++ backend/src/Designer/Hubs/Sync/SyncError.cs
@@ -1,4 +1,4 @@
-namespace Altinn.Studio.Designer.Hubs.SyncHub;
+namespace Altinn.Studio.Designer.Hubs.Sync;
 
 public record SyncError(string ErrorCode, Source Source, string Details);
