+++ backend/src/Designer/Hubs/Sync/SyncErrorCodes.cs
@@ -1,4 +1,4 @@
-namespace Altinn.Studio.Designer.Hubs.SyncHub;
+namespace Altinn.Studio.Designer.Hubs.Sync;
 
 public static class SyncErrorCodes
 {
