+++ backend/src/Designer/Hubs/Sync/SyncSuccess.cs
@@ -1,0 +1,3 @@
+namespace Altinn.Studio.Designer.Hubs.Sync;
+
+public record SyncSuccess(Source Source);
