--- backend/src/Designer/Hubs/SyncHub/SyncSuccess.cs
@@ -1,3 +1,0 @@
-namespace Altinn.Studio.Designer.Hubs.SyncHub;
-
-public record SyncSuccess(Source Source);
